# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: 匿名树洞社区 API
  version: 1.0.0
  description: 匿名发帖/评论、话题标签、点赞热度、每日精选、敏感词审核
servers:
  - url: http://localhost:19401
paths:
...
  /api/v1/posts:
    get:
      summary: 最新/标签/精选帖子列表
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
        - in: query
          name: tag_id
          schema: { type: integer }
        - in: query
          name: featured
          schema: { type: boolean }
      responses:
        '200':
          description: 分页帖子列表
    post:
      summary: 发布帖子
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                title: { type: string }
                content: { type: string }
                images: { type: array, items: { type: string } }
                tags: { type: array, items: { type: string } }
      responses:
        '200':
          description: 帖子与命中信息
...
  /api/v1/posts/{id}:
    get:
      summary: 帖子详情
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 帖子详情
  /api/v1/posts/{id}/favorite:
    post:
      summary: 收藏帖子（幂等，重复收藏不产生重复记录）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 收藏结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  favorited: { type: boolean }
                  favoriteCount: { type: integer }
    delete:
      summary: 取消收藏（幂等）
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 收藏结果
  /api/v1/favorites:
    get:
      summary: 我的收藏（按最近收藏时间倒序，被审核屏蔽的帖子自动隐藏、放行后重新出现）
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        '200':
          description: 分页帖子列表
...